+++ customer-dashboard/src/main/java/com/targetindia/dao/CustomerDao.java
@@ -8,11 +8,11 @@ public interface CustomerDao {
     // CRUD operations related to Customer
     public void addCustomer(Customer customer) throws DaoException;
 
-    public Customer findById(int id) throws DaoException;
+    public Customer findById(long id) throws DaoException;
 
     public void updateCustomer(Customer customer) throws DaoException;
 
-    public void deleteCustomer(int id) throws DaoException;
+    public void deleteCustomer(long id) throws DaoException;
 
     // Query operations related to Customer
